Shift application creation to factory method for testing
