--- feeder/templates/fed.html
@@ -1,1 +1,0 @@
-<p>Sent feed request to {{gateway}}</p>
