--- feeder/templates/welcome.html
@@ -1,59 +1,0 @@
-<p>List of all discovered devices:</p>
-<ul>
-    {% for device in devices %}
-        <li>
-            <h3>Device: {{device["name"]}}</h3>
-            <h4>HID: {{device["hid"]}}</h4>
-        <form data-command="feed" data-gateway-id="{{device["gatewayHid"]}}" data-device-id="{{device["hid"]}}">
-            <label for="portion">Portion:</label>
-            <input type="text" name="portion" value="0.0625"/>
-            <button type="button">Feed</button>
-        </form>
-        <form data-command="button" data-gateway-id="{{device["gatewayHid"]}}" data-device-id="{{device["hid"]}}">
-            <label for="enable">Enable:</label>
-            <input type="checkbox" name="enable"/>
-            <button type="button">Set Front Button</button>
-        </form>
-        <form data-command="utc_offset" data-gateway-id="{{device["gatewayHid"]}}" data-device-id="{{device["hid"]}}">
-            <label for="utc_offset">
-            <input type="number" name="utc_offset" value="-7"/>
-            <button type="button">Set UTC Offset</button>
-        </form>
-        <form data-command="reboot" data-gateway-id="{{device["gatewayHid"]}}" data-device-id="{{device["hid"]}}">
-            <button type="button">Reboot</button>
-        </form>
-    </li>
-    {% endfor %}
-</ul>
-<script>
-    function serializeArray(form) {
-        var objects = [];
-        if (typeof form == 'object' && form.nodeName.toLowerCase() == "form") {
-            var fields = form.getElementsByTagName("input");
-            for(var i = 0; i < fields.length; i++) {
-                objects[objects.length] = { name: fields[i].getAttribute("name"), value: fields[i].getAttribute("value") };
-            }
-        }
-        return objects;
-    }
-
-    function sendCommand(command, gateway_id, device_id, form) {
-        var apiUrl = "/api/v1/feeder/" + gateway_id + "/" + device_id + "/" + command;
-        var dataObject = serializeArray(form);
-
-        fetch(apiUrl, {
-          method: "POST",
-          body: JSON.stringify(dataObject)
-        }).then(res => {
-          console.log("Request complete! response:", res);
-        });
-    }
-
-    var forms = document.getElementsByTagName('form');
-    for (var i = 0; i < forms.length; i++) {
-        var button = forms[i].getElementsByTagName('button')[0];
-        button.addEventListener('click', function() {
-            sendCommand(this.getAttribute('data-command'), this.getAttribute('data-gateway-id'), this.getAttribute('data-device-id'), this);
-        }.bind(forms[i]));
-    }
-</script>
